[Feature] Define a list of software to only update
There are some programs for which I keep copies of their Git
repositories on my computer, and I want to keep those up to date all
the time.  This patch creates a new list for those programs and adds
an additional `(for-each ...)` to the main logic to call `update` on
those programs.  This change allows me to remove a few programs from
the section of code that defines software with special build and
install functions, because some had no-ops for those two functions
since all I wanted to do was update their repositories.  By putting
them in this new, secondary list of software I don't need to create
no-op functions for them.