[Feature] Define a generic function for updating Git repositories
This patch implements `git-update`, a function which performs the
following steps:

1. It runs git-fetch on a given remote.

2. It runs git-checkout for a given branch.

3. It runs git-merge --ff-only on the remote version of that branch.

The implementation uses `case-lambda` to support variable arity for
the `git-update` function.  When called with no arguments it runs

    (git-update "origin" "master")

which is ideal for most repositories.  But putting this logic into a
separate function makes it possible later to more easily write custom
update functions for software that needs to either pull from a
different remote or update a different branch, or both.

See-Also: http://srfi.schemers.org/srfi-16/