[Refactor] Put all software to process into a single list
This patch combines `important-software` and `software-to-update-only`
into a single list: `software-to-process`.  The immediate benefit of
this change is that it allows for some simplification of the logic in
the main procedure.  It also opens the door to implementing a feature
by which we can tell the script to only process specific software via
a command-line flag, because that will be easier to implement if all
of the software is contained within a single list instead of having to
check two (or more) for the data for said software.