[Feature] Add a function for finding a specific :software record
This patch implements `(find-software name)` which will search through
`software-to-process` and return the record for the software which
matches the given name.  The search is case-insensitive, so for
example `(find-software "php")` will return the record for `"PHP"`.

The function actually returns a list of matches even though for now
there should only ever be one match.  But in the future we may support
processing multiple versions of software with the same name, e.g. to
install stable and bleeding-edge versions of the same program.

See-Also: http://srfi.schemers.org/srfi-1/srfi-1.html