[Feature] Implement `get-name` function for :software records
Within the `process` function we create a name for each :software
record by tokenizing and manipulating the directory path for that
record.  This patch not only creates a new function to create that
name, but also simplifies the logic.  We do not need to make a
character set with which to tokenize the directory.  Instead we can
simply use the `file-name-nondirectory` function that is a standard
part of the Scheme Shell, giving us the same result as our manual
tokenizing code.