[Refactor] Clean up the duplicated logic in processor functions
In this context "processor functions" refers specifically to:

1. `update`
2. `build`
3. `install`

All three functions have similar, duplicated logic:

    (with-cwd (software)
        (or (and-let* [Try running custom behavior])
            [Run default behavior]))

"Custom behavior" is that defined by :software records which provide
procedures for use with `get-update-function` et al.  For each
processor we want to see if the given record has such a procedure
implementing any custom behavior.  If so then we call that.  Otherwise
we fall back on default logic such as calling `git-update`.

This patch refactors the code by wrapping the logic described above
into a utility called `create-processor`.  It lets us simplify the
definition for such processors as `update` by writing it as:

    (define update (create-processor get-update-function git-update))

In this case, `update` becomes a function that accepts a :software
record (as usual) and will first call `get-update-function` on the
given record.  If `get-update-function` returns a procedure then
`update` calls that procedure on the record.  Otherwise `update` calls
`git-update` on the record.

The patch rewrites `update`, `build`, and `install` in terms of
`create-processor`, which simplifies them greatly.