[Fix] Problems with string functions on :software records
This reverts commit 0547c050 ("[Refactor] Clean up the duplicated
logic in processor functions").  That commit created two bugs related
to "stringifying" `:software` records and calling `string-length` on
their properties.  This patch fixes those problems by reverting the
commit which introduced them, although this leads to an increase in
the code due to re-introducing the duplication of code which commit
0547c050 refactored and removed.