[Feature] Add a name property to all software records
Currently we obtain the name of software by extracting it from the
`directory` property of the :software record.  However, this makes it
difficult to implement some functionality, particularly support for
updating software via package managers; we want to call those package
managers with only a name, e.g. `pip install --upgrade foobar`, and we
can only provide that name by giving a bogus value to the `directory`
property which ends with the name we want.

This patch addresses the problem by adding an explicit `name` property
to all :software records.  That allows us to create records intended
for use with package managers where all we want to do is provide a
name and not care about the directory (the package manager will know
the directory to use).  This requires refactoring the `get-name`
function, which is now a getter provided by the :software record
definition, the `find-software` function, and all uses of
`new-software` to account for the new `name` parameter.